Address UI customisation feedback
